refactor: position 타입 long에서 int로 변경
문자열 반복 메서드인 repeat은 int 타입을 인자로 받고 있는데, position은 long 타입이라 변경하였습니다.

+++ src/main/java/org/duckstudy/movingcar/Car.java
@@ -3,13 +3,13 @@ package org.duckstudy.movingcar;
 public class Car {
 
     private final String name;
-    private long position;
+    private int position;
 
     public String getName() {
         return name;
     }
 
-    public long getPosition() {
+    public int getPosition() {
         return position;
     }
 
@@ -18,7 +18,7 @@ public class Car {
         this.position = 0;
     }
 
-    public long move() {
+    public int move() {
         return ++position;
     }
